add useLastCommissionChange hook to its own file
